before majot refactoring

+++ backend/boards/boardsController.js
@@ -5,23 +5,32 @@ GET
   route :  /
 */
 
-const getBoards = (req, res) => {
-  Board.find({ adminUser: req.user._id })
-    .then(board => {
-      return res.json(board)
-    })
-    .catch(err => res.status(400).json('Error: ' + err))
-}
+// const getBoards = async (req, res) => {
+//   try {
+//     const board = await Board.find({ adminUser: req.user._id })
+//     return res.status(200).json(board)
+//   } catch (err) {
+//     return res
+//       .status(500)
+//       .json({ message: 'There was an error. Please try again later' })
+//   }
+// }
 
 const getBoardsTeam = async (req, res) => {
-  const boards = await Board.find()
-  const teamBoards = []
-  boards.forEach(board => {
-    if (board.team.includes(req.user._id)) {
-      teamBoards.push(board)
-    }
-  })
-  return res.json(teamBoards)
+  try {
+    const boards = await Board.find()
+    const teamBoards = []
+    boards.forEach(board => {
+      if (board.team.includes(req.user._id)) {
+        teamBoards.push(board)
+      }
+    })
+    return res.status(200).json({ boards: teamBoards, user: req.user })
+  } catch (err) {
+    return res
+      .status(500)
+      .json({ message: 'There was an error. Please try again later' })
+  }
 }
 
 /*
@@ -29,15 +38,19 @@ const getBoardsTeam = async (req, res) => {
   route :  /
 */
 
-const createBoard = (req, res) => {
-  const board = new Board()
-  board.boardName = req.body.boardName
-  board.adminUser = req.user._id
-  board.team.push(req.user._id)
-  board
-    .save()
-    .then(() => res.json({ boardId: board._id }))
-    .catch(err => res.status(400).json('Error: ' + err))
+const createBoard = async (req, res) => {
+  try {
+    const board = new Board()
+    board.boardName = req.body.boardName
+    board.adminUser = req.user._id
+    board.team.push(req.user._id)
+    await board.save()
+    res.status(201).json({ boardId: board._id })
+  } catch (err) {
+    return res
+      .status(500)
+      .json({ message: 'There was an error. Please try again later' })
+  }
 }
 
 /*
@@ -46,16 +59,20 @@ const createBoard = (req, res) => {
   id is the BoardId
 */
 
-const updateBoard = (req, res) => {
-  Board.findById(req.params.id)
-    .then(board => {
+const updateBoard = async (req, res) => {
+  try {
+    const board = await Board.findById(req.params.id)
+    if (board.team.includes(req.user._id)) {
       board[`${req.body.name}`] = req.body.value
-      board
-        .save()
-        .then(() => res.json('updated'))
-        .catch(err => res.json('ERROR: ' + err))
-    })
-    .catch(err => res.status(400).json('Error: ' + err))
+      await board.save()
+      return res.status(200).json({ msg: 'updated' })
+    }
+    return res.status(400).json({ msg: 'not authorised to update this board' })
+  } catch (err) {
+    return res
+      .status(500)
+      .json({ message: 'There was an error. Please try again later' })
+  }
 }
 
 /*
@@ -64,10 +81,19 @@ const updateBoard = (req, res) => {
   id is the BoardID
 */
 
-const deleteBoard = (req, res) => {
-  Board.findByIdAndDelete(req.params.id)
-    .then(() => res.json('deleted Board'))
-    .catch(err => res.status(400).json('Error: ' + err))
+const deleteBoard = async (req, res) => {
+  try {
+    const board = await Board.findById(req.params.id)
+    if (req.user._id === board.adminUser.toString()) {
+      await Board.findByIdAndDelete(req.params.id)
+      return res.status(200).json({ msg: 'deleted Board' })
+    }
+    res.status(400).json({ msg: 'not authorised to delete this board' })
+  } catch (err) {
+    return res
+      .status(500)
+      .json({ msg: 'There was an error. Please try again later' })
+  }
 }
 
 /*
@@ -77,10 +103,19 @@ const deleteBoard = (req, res) => {
 */
 
 const addTeamMember = async (req, res) => {
-  const board = await Board.findById(req.params.id)
-  board.team.push(req.body.teamMemberId)
-  await board.save()
-  res.json(board)
+  try {
+    const board = await Board.findById(req.params.id)
+    if (board.team.includes(req.user._id)) {
+      board.team.push(req.body.teamMemberId)
+      await board.save()
+      return res.status(200).json(board)
+    }
+    return res.status(400).json({ msg: 'not authorised to delete this board' })
+  } catch (err) {
+    return res
+      .status(500)
+      .json({ msg: 'There was an error. Please try again later' })
+  }
 }
 
 /*
@@ -89,16 +124,25 @@ const addTeamMember = async (req, res) => {
   id is the boardId
 */
 const removeTeamMember = async (req, res) => {
-  const board = await Board.findById(req.params.id)
-  board.team = board.team.filter(
-    teamMemberId => teamMemberId != req.body.teamMemberId
-  )
-  await board.save()
-  res.json(board)
+  try {
+    const board = await Board.findById(req.params.id)
+    if (board.team.includes(req.user._id)) {
+      board.team = board.team.filter(
+        teamMemberId => teamMemberId.toString() !== req.body.teamMemberId
+      )
+      await board.save()
+      return res.status(200).json(board)
+    }
+    return res.status(400).json({ msg: 'not authorised to delete this board' })
+  } catch (err) {
+    return res
+      .status(500)
+      .json({ msg: 'There was an error. Please try again later' })
+  }
 }
 
 module.exports = {
-  getBoards,
+  // getBoards,
   getBoardsTeam,
   createBoard,
   deleteBoard,
